+++ app/client/event_listeners/event_listeners.js
@@ -29,6 +29,7 @@ import { listener_searches_save } from "./listener_searches_save.js";
 import { listener_searches_overwrite } from "./listener_searches_overwrite.js";
 import { listener_ratings_crimes } from "./listener_ratings_crimes.js";
 import { listener_ratings_renters } from "./listener_ratings_renters.js";
+import { listener_toggle_intro_modal } from "./listener_toggle_intro_modal.js";
 
 import { filter_monochrome } from "./listener_filter_monochrome.js";
 import { filter_invert } from "./listener_filter_invert.js";
@@ -61,6 +62,7 @@ export function load_event_listeners(){
     listener_searches_overwrite()
     listener_ratings_crimes()
     listener_ratings_renters()
+    listener_toggle_intro_modal()
     
     filter_monochrome()
     filter_invert()