+++ app/client/event_listeners/listener_toggle_intro_modal.js
@@ -1,0 +1,23 @@
+// ┌─────────────────────────────────────┐
+// │                                     │
+// │          Toggle Intro Modal         │
+// │                                     │
+// └─────────────────────────────────────┘
+
+export function listener_toggle_intro_modal(){
+
+    const toggleIntroModalSwitch = document.getElementById('introModalToggle');
+
+    toggleIntroModalSwitch.addEventListener('click', async function(event) {
+
+        console.log(this.checked)
+
+        if ( this.checked ){
+            window.localStorage.setItem('introModalToggle', 'true');
+        } else {
+            window.localStorage.setItem('introModalToggle', 'false');
+        }
+
+    });
+
+}