+++ app/client/intro/intro_modal.js
@@ -1,7 +1,15 @@
 
 export function intro_modal()
 {
+    // Get current state of toggle
+    let introModalToggleState = window.localStorage.getItem('introModalToggle');
 
+    // Set the toggle html 'checked' to be the same as the value in localstorage
+    if (introModalToggleState == 'true')
+    {
+        document.getElementById('introModalToggle').checked = true;
+    }
+    
     // set the modal menu element
     const targetEl = document.getElementById('popup-modal-intro');
     
@@ -11,22 +19,22 @@ export function intro_modal()
         placement: 'center',
         backdropClasses: 'bg-gray-900 bg-opacity-50 dark:bg-opacity-80 fixed inset-0 z-40',
         onHide: () => {
-            console.log('modal is hidden');
         },
         onShow: () => {
-            console.log('modal is shown');
         },
         onToggle: () => {
-            console.log('modal has been toggled');
         }
     };
 
     var introModal = new Modal(targetEl, options);
 
-    // show the modal
-    introModal.show();
+    console.log('introModalToggleState: ' + introModalToggleState)
 
-
+    if (introModalToggleState == 'false' || introModalToggleState == null)
+    {
+        // show the modal
+        introModal.show();
+    }
 
     // Close the Modal button
     const popupModalIntroClose = document.getElementById('popup-modal-intro-close');